requirements.txt
chromadb==0.5.23
google-generativeai==0.7.2
langchain==0.2.16
langchain-chroma==0.1.4
langchain-community==0.2.16
langchain-core==0.2.38
langchain-google-genai==1.0.10
pypdf==6.7.1
python-dotenv==1.2.1
